master/requirements.txt
flask==3.0.0
redis==5.0.1
mysql-connector-python==8.0.33